requirements.txt
langchain
google-generativeai
streamlit
langchain-google-genai
python-dotenv
chromadb
pypdf
PyPDF2
transformers
